Move snapshot validations to validation-webhook directory
