+++ pkg/utils/util.go
@@ -25,7 +25,6 @@ import (
 	"time"
 
 	crdv1 "github.com/kubernetes-csi/external-snapshotter/client/v4/apis/volumesnapshot/v1"
-	crdv1beta1 "github.com/kubernetes-csi/external-snapshotter/client/v4/apis/volumesnapshot/v1beta1"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -120,87 +119,6 @@ var SnapshotterListSecretParams = secretParamsMap{
 	name:               "SnapshotterList",
 	secretNameKey:      PrefixedSnapshotterListSecretNameKey,
 	secretNamespaceKey: PrefixedSnapshotterListSecretNamespaceKey,
-}
-
-// ValidateV1Snapshot performs additional strict validation.
-// Do NOT rely on this function to fully validate snapshot objects.
-// This function will only check the additional rules provided by the webhook.
-func ValidateV1Snapshot(snapshot *crdv1.VolumeSnapshot) error {
-	if snapshot == nil {
-		return fmt.Errorf("VolumeSnapshot is nil")
-	}
-
-	vscname := snapshot.Spec.VolumeSnapshotClassName
-	if vscname != nil && *vscname == "" {
-		return fmt.Errorf("Spec.VolumeSnapshotClassName must not be the empty string")
-	}
-	return nil
-}
-
-// ValidateV1Beta1Snapshot performs additional strict validation.
-// Do NOT rely on this function to fully validate snapshot objects.
-// This function will only check the additional rules provided by the webhook.
-func ValidateV1Beta1Snapshot(snapshot *crdv1beta1.VolumeSnapshot) error {
-	if snapshot == nil {
-		return fmt.Errorf("VolumeSnapshot is nil")
-	}
-
-	source := snapshot.Spec.Source
-
-	if source.PersistentVolumeClaimName != nil && source.VolumeSnapshotContentName != nil {
-		return fmt.Errorf("only one of Spec.Source.PersistentVolumeClaimName = %s and Spec.Source.VolumeSnapshotContentName = %s should be set", *source.PersistentVolumeClaimName, *source.VolumeSnapshotContentName)
-	}
-	if source.PersistentVolumeClaimName == nil && source.VolumeSnapshotContentName == nil {
-		return fmt.Errorf("one of Spec.Source.PersistentVolumeClaimName and Spec.Source.VolumeSnapshotContentName should be set")
-	}
-	vscname := snapshot.Spec.VolumeSnapshotClassName
-	if vscname != nil && *vscname == "" {
-		return fmt.Errorf("Spec.VolumeSnapshotClassName must not be the empty string")
-	}
-	return nil
-}
-
-// ValidateV1SnapshotContent performs additional strict validation.
-// Do NOT rely on this function to fully validate snapshot content objects.
-// This function will only check the additional rules provided by the webhook.
-func ValidateV1SnapshotContent(snapcontent *crdv1.VolumeSnapshotContent) error {
-	if snapcontent == nil {
-		return fmt.Errorf("VolumeSnapshotContent is nil")
-	}
-
-	vsref := snapcontent.Spec.VolumeSnapshotRef
-
-	if vsref.Name == "" || vsref.Namespace == "" {
-		return fmt.Errorf("both Spec.VolumeSnapshotRef.Name = %s and Spec.VolumeSnapshotRef.Namespace = %s must be set", vsref.Name, vsref.Namespace)
-	}
-
-	return nil
-}
-
-// ValidateV1Beta1SnapshotContent performs additional strict validation.
-// Do NOT rely on this function to fully validate snapshot content objects.
-// This function will only check the additional rules provided by the webhook.
-func ValidateV1Beta1SnapshotContent(snapcontent *crdv1beta1.VolumeSnapshotContent) error {
-	if snapcontent == nil {
-		return fmt.Errorf("VolumeSnapshotContent is nil")
-	}
-
-	source := snapcontent.Spec.Source
-
-	if source.VolumeHandle != nil && source.SnapshotHandle != nil {
-		return fmt.Errorf("only one of Spec.Source.VolumeHandle = %s and Spec.Source.SnapshotHandle = %s should be set", *source.VolumeHandle, *source.SnapshotHandle)
-	}
-	if source.VolumeHandle == nil && source.SnapshotHandle == nil {
-		return fmt.Errorf("one of Spec.Source.VolumeHandle and Spec.Source.SnapshotHandle should be set")
-	}
-
-	vsref := snapcontent.Spec.VolumeSnapshotRef
-
-	if vsref.Name == "" || vsref.Namespace == "" {
-		return fmt.Errorf("both Spec.VolumeSnapshotRef.Name = %s and Spec.VolumeSnapshotRef.Namespace = %s must be set", vsref.Name, vsref.Namespace)
-	}
-
-	return nil
 }
 
 // MapContainsKey checks if a given map of string to string contains the provided string.
